app/requirements.txt
asgiref==3.11.0
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.4.0
django-axes==8.1.0
django-cors-headers==4.9.0
django-password-validators==1.7.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
reportlab==4.2.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
